+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will now be documented in this file.
 
+## [2.0.7] - 2019-12-12
+- Removed unused variables from config 
+- Pass Address Postcode as sanity value
+- Widget Button Text
+- Widget Footer
+- Widget Mode option
+
 ## [2.0.6] - 2019-12-04
 - Specified Country Logic Added
 - Rounding Logic corrected